src/requirements-dev.txt
# Development Requirements for CivicForge

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0

# Code Quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
isort>=5.12.0

# Core Dependencies
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6
pyyaml>=6.0.0

# NLP (start simple)
spacy>=3.5.0
textblob>=0.17.1
sentence-transformers>=2.2.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Database
sqlalchemy>=2.0.0
alembic>=1.11.0

# Development Tools
ipython>=8.12.0
pre-commit>=3.3.0

# Documentation
mkdocs>=1.4.0
mkdocs-material>=9.1.0